Release/1.1.3 to main

+++ core/data/src/main/java/com/yapp/core/data/remote/model/response/AttendStatisticsResponse.kt
@@ -1,14 +1,13 @@
 package com.yapp.core.data.remote.model.response
 
 import com.yapp.model.AttendStatistics
-import com.yapp.model.UserInfo
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class AttendStatisticsResponse(
     val totalSessionCount: Int,
     val remainingSessionCount: Int,
-    val sessionProgressRate: Int,
+    val sessionProgressRate: Double,
     val attendancePoint: Int,
     val attendanceCount: Int,
     val lateCount: Int,
