+++ core/model/src/main/java/com/yapp/model/AttendStatistics.kt
@@ -3,7 +3,7 @@ package com.yapp.model
 data class AttendStatistics(
     val totalSessionCount: Int,
     val remainingSessionCount: Int,
-    val sessionProgressRate: Int,
+    val sessionProgressRate: Double,
     val attendancePoint: Int,
     val attendanceCount: Int,
     val lateCount: Int,
